Add files via upload
The following project provided in this folder is a first year python project. The intent of this project was simplicity, but I decided that I should try to go above and beyond to really push my skills as a programmer and my patients with debugging issues. Following this, the theme of the game is Dante's Inferno, the game features a loading system that converts from serialized data to readable text. The loading animation has been modified by myself and credited as to where credit is due.
Feel free to use this to your heart's content. 
-Brandon P.